Add HMR for CSV files
