+++ vite.config.js
@@ -3,7 +3,7 @@ import react from '@vitejs/plugin-react';
 import dsv from '@rollup/plugin-dsv';
 
 import nunjucks from './vite-plugin-transform-nunjucks';
-import hmrjson from './vite-plugin-hmr-json';
+import hmr from './vite-plugin-hmr';
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -11,6 +11,6 @@ export default defineConfig({
     react(),
     dsv(),
     nunjucks(),
-    hmrjson(),
+    hmr(),
   ]
 })
